Make new bookmarks appear immedietly
As soon as a bookmark is added, it will be moved to the front.
All this took was one extra character to fix a variable name from earlier in development, but my editing program decided to reformat some empty space, making it look like I changed a lot, when in reality I only changed one line.
I think we still could clean up the code a bit, and comments will help if we need to fix a bug at some point in the future.